+++ Client/Frontend/Settings/Main/Support/OpenSupportPageSetting.swift
@@ -1,0 +1,23 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+import Shared
+
+/// Opens the SUMO page in a new tab
+class OpenSupportPageSetting: Setting {
+    init(delegate: SettingsDelegate?, theme: Theme) {
+        super.init(title: NSAttributedString(string: .AppSettingsHelp,
+                                             attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary]),
+                   delegate: delegate)
+    }
+
+    override func onClick(_ navigationController: UINavigationController?) {
+        navigationController?.dismiss(animated: true) {
+            if let url = URL(string: "https://support.mozilla.org/products/ios") {
+                self.delegate?.settingsOpenURLInNewTab(url)
+            }
+        }
+    }
+}
